I#22 move span under cpp subdirectory
