+++ include/toolbox/cpp/cstring_ref.hpp
@@ -3,7 +3,7 @@
 
 #include <cstring>
 #include <string>
-#include "toolbox/span.hpp"
+#include "toolbox/cpp/span.hpp"
 #include "toolbox/std/type_traits.hpp"
 
 namespace toolbox
